pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sandbox"
version = "0.1.0"
description = "Military sandbox roundtable discussion engine powered by AIHubMix-compatible APIs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
  "python-dotenv>=1.0,<2.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.scripts]
sandbox-desktop = "sandbox.desktop.main:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
]
desktop = [
  "PySide6>=6.7,<7.0",
]

[tool.setuptools.packages.find]
include = ["sandbox*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
